Lab10.js
In this sketch, I made a sample frogger style game using classes to build the lanes and the cars bouncing across the screen. In order to play the game, you must click on the canvas and use the movement keys to make the frog move up the screen. Note: that the frog is off screen until you start playing the game. Click play to restart the game and the positions of the cars.